todo-list-backend-django/todolist/requirements.txt
Django>=4.0,<5.0  # Core framework  
djangorestframework  # For creating REST APIs  
psycopg2-binary  # PostgreSQL database adapter for Python  
gunicorn  # WSGI HTTP server for UNIX  
